Enhancement / Fix: Implement Inspect command + fixes on infra deployment (#18)
* build command setup

* cleanup and refactors

* remove explorer as artifact plus setup and fixes

* terraform values as var

* cleanups

* fixes and cleanup. runner and deployer implementations

* functional dev command

* dev command working

* deploy implementation

* deploy terraform fixes

* quick inspect infra

* logs and cleanup

* fixes

* fix and cleanup

* cleanup

* cleanup

* fix contracts artifacts zip

* small fixes

* terraform extend timeout and increase nodes. Add dependencies, quiet command and ingress instal on dev

* install cert manager

* fix terraform and implement inspect command

* fixes

* expand readme

* fix merge

* cleanup and improvements

+++ opraas_core/src/utils/system.rs
@@ -29,6 +29,10 @@ pub fn execute_command(command: &mut Command, silent: bool) -> Result<String, St
 }
 
 pub fn copy_and_overwrite(src: &PathBuf, dest: &PathBuf) -> io::Result<()> {
+    if src == dest {
+        return Ok(());
+    }
+
     if dest.exists() {
         fs::remove_file(dest)?;
     }
